We can now write evlrs using laszip backend
And reactivate the corresponding tests
