+++ pylas/lasappender.py
@@ -99,13 +99,13 @@ class LazrsAppender:
 
 class LasAppender:
     def __init__(
-            self,
-            dest: BinaryIO,
-            laz_backend: Union[LazBackend, Iterable[LazBackend]] = (
-                    LazBackend.LazrsParallel,
-                    LazBackend.Lazrs,
-            ),
-            closefd: bool = True,
+        self,
+        dest: BinaryIO,
+        laz_backend: Union[LazBackend, Iterable[LazBackend]] = (
+            LazBackend.LazrsParallel,
+            LazBackend.Lazrs,
+        ),
+        closefd: bool = True,
     ) -> None:
         if not dest.seekable():
             raise TypeError("Expected the 'dest' to be a seekable file object")
@@ -126,7 +126,7 @@ class LasAppender:
 
         if header.version.minor >= 4 and header.number_of_evlrs > 0:
             assert (
-                    self.dest.tell() <= self.header.start_of_first_evlr
+                self.dest.tell() <= self.header.start_of_first_evlr
             ), "The position is past the start of evlrs"
             pos = self.dest.tell()
             self.dest.seek(self.header.start_of_first_evlr, io.SEEK_SET)
@@ -156,7 +156,11 @@ class LasAppender:
             self.dest.close()
 
     def _write_evlrs(self) -> None:
-        if self.header.version.minor >= 4 and self.evlrs is not None and len(self.evlrs) > 0:
+        if (
+            self.header.version.minor >= 4
+            and self.evlrs is not None
+            and len(self.evlrs) > 0
+        ):
             self.header.number_of_evlr = len(self.evlrs)
             self.header.start_of_first_evlr = self.dest.tell()
             self.evlrs.write_to(self.dest, as_extended=True)
@@ -168,11 +172,11 @@ class LasAppender:
         self.dest.seek(pos, io.SEEK_SET)
 
     def _create_laz_backend(
-            self,
-            laz_backend: Union[LazBackend, Iterable[LazBackend]] = (
-                    LazBackend.LazrsParallel,
-                    LazBackend.Lazrs,
-            ),
+        self,
+        laz_backend: Union[LazBackend, Iterable[LazBackend]] = (
+            LazBackend.LazrsParallel,
+            LazBackend.Lazrs,
+        ),
     ) -> LazrsAppender:
         try:
             laz_backend = iter(laz_backend)
