+++ pylas/vlrs/vlr.py
@@ -75,5 +75,3 @@ class VLR(BaseVLR):
         return "<{}(user_id: '{}', record_id: '{}', data len: {})>".format(
             self.__class__.__name__, self.user_id, self.record_id, len(self.record_data)
         )
-
-
